add no_metadata_cache argument to copernicusmarine_credentials() (#787)
* add no_metadata_cache argument to copernicusmarine_credentials() to support copernicusmarine>=1.0.2

* updated whatsnew

* modularity in copernicusmarine_reset()

* update minimal cmc version

+++ dfm_tools/download.py
@@ -316,26 +316,29 @@ def copernicusmarine_credentials():
         InvalidUsernameOrPassword,
         get_and_check_username_password,
     )
+    login_kwargs = dict(username=None, password=None, credentials_file=DEFAULT_CLIENT_CREDENTIALS_FILEPATH, no_metadata_cache=False)
     if not DEFAULT_CLIENT_CREDENTIALS_FILEPATH.is_file():
         print("Downloading CMEMS data requires a Copernicus Marine username and password, sign up for free at: https://data.marine.copernicus.eu/register.")
-        username, password = get_and_check_username_password(username=None, password=None, credentials_file=DEFAULT_CLIENT_CREDENTIALS_FILEPATH)
+        username, password = get_and_check_username_password(**login_kwargs)
         success = copernicusmarine.login(username, password)
         if not success:
             raise InvalidUsernameOrPassword("invalid credentials")
     else:
-        username, password = get_and_check_username_password(username=None, password=None, credentials_file=DEFAULT_CLIENT_CREDENTIALS_FILEPATH)
+        username, password = get_and_check_username_password(**login_kwargs)
     return username, password
 
 
-def copernicusmarine_reset():
-    print("resetting Copernicus Marine Toolbox, you will have to login again.")
-    dir_copernicusmarine = os.path.expanduser("~/.copernicusmarine")
-    print(f"- removing {dir_copernicusmarine}")
-    shutil.rmtree(dir_copernicusmarine, ignore_errors=True)
-    print("- overwriting copernicusmarine metadata cache (takes some time)")
-    subprocess.run("copernicusmarine describe --overwrite-metadata-cache")
-    print("- updating copernicusmarine")
-    subprocess.run("pip install copernicusmarine -U")
+def copernicusmarine_reset(remove_folder=False, overwrite_cache=True, update_package=False):
+    if remove_folder:
+        dir_copernicusmarine = os.path.expanduser("~/.copernicusmarine")
+        print(f"reset copernicusmarine: removing {dir_copernicusmarine}, you will have to login again.")
+        shutil.rmtree(dir_copernicusmarine, ignore_errors=True)
+    if overwrite_cache:
+        print("reset copernicusmarine: overwriting copernicusmarine metadata cache (takes some time)")
+        subprocess.run("copernicusmarine describe --overwrite-metadata-cache")
+    if update_package:
+        print("reset copernicusmarine: updating copernicusmarine")
+        subprocess.run("pip install copernicusmarine -U")
 
 
 def copernicusmarine_dataset_timerange(dataset_id):
